Split coop TUI completion rendering
Committed-By-Agent: codex
